PostView2: Updated PostView2 project.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@12542 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostView2/4PointAngleTool.cpp
@@ -167,7 +167,7 @@ void C4PointAngleTool::update(bool breset)
 		{
 			FEMeshBase& mesh = *doc->GetActiveMesh();
 			const vector<FENode*> selectedNodes = doc->GetGLModel()->GetNodeSelection();
-			int N = selectedNodes.size();
+			int N = (int)selectedNodes.size();
 			int nsel = 0;
 			for (int i = 0; i<N; ++i)
 			{
